refactor: Plugin and Plugins file names changed
