+++ src/components/Modal/Plugins.vue
@@ -1,98 +1,106 @@
 <template>
   <UiModal :open="open" @close="$emit('close')">
     <template v-slot:header>
-      <h3>{{ $t('plugins') }}</h3>
+      <h3>
+        {{ plugin.name ? $t('settings.editPlugin') : $t('settings.addPlugin') }}
+      </h3>
     </template>
-    <div class="m-4 mt-4" v-if="selected === false">
-      <div
-        v-for="(plugin, i) in plugins"
-        :key="i"
-        class="mb-3 p-4 border rounded-2 text-white text-center"
-      >
-        <img
-          class="circle border"
-          :src="getLogoUrl(i)"
-          width="64"
-          height="64"
-        />
-        <h3 v-text="plugin.name" />
-        <div class="mb-2">
-          <a :href="plugin.website" target="_blank" class="text-white">
-            {{ $t('learnMore') }}
-            <Icon name="external-link" />
-          </a>
-        </div>
-        <UiButton @click="selected = i">
-          {{ !form[i] ? $t('add') : $t('edit') }}
-        </UiButton>
-      </div>
-    </div>
-    <template v-if="selected === false" v-slot:footer>
-      <div class="col-6 float-left pr-2">
-        <UiButton @click="$emit('close')" type="button" class="width-full">
-          {{ $t('cancel') }}
-        </UiButton>
-      </div>
-      <div class="col-6 float-left pl-2">
+    <Search
+      v-if="!plugin.name && !input.name"
+      v-model="searchInput"
+      :placeholder="$t('searchPlaceholder')"
+      :modal="true"
+    />
+    <div class="mt-4 mx-0 mx-md-4">
+      <div v-if="input.name" class="p-4 mb-4 border rounded-2 text-white">
+        <h4 v-text="input.name" class="mb-3 text-center" />
         <UiButton
-          @click="[$emit('update:modelValue', form), $emit('close')]"
-          type="submit"
-          class="width-full button--submit"
+          class="d-block width-full mb-3 overflow-x-auto"
+          style="height: auto;"
         >
-          {{ $t('save') }}
+          <TextareaAutosize
+            v-model="input.params"
+            :placeholder="$t('settings.pluginParameters')"
+            class="input text-left"
+            style="width: 560px;"
+          />
+        </UiButton>
+        <UiButton
+          @click="handleSubmit"
+          :disabled="!isValid"
+          class="button--submit width-full"
+        >
+          {{ plugin.name ? $t('save') : $t('add') }}
         </UiButton>
       </div>
-    </template>
-    <div v-if="selected !== false" class="m-4 p-4 border rounded-2 text-white">
-      <PluginAragonConfig
-        :proposal="proposal"
-        v-model="form.aragon"
-        @close="selected = false"
-        v-if="selected === 'aragon'"
-      />
-      <PluginGnosisConfig
-        :proposal="proposal"
-        :network="space.network"
-        v-model="form.gnosis"
-        @close="selected = false"
-        v-if="selected === 'gnosis'"
-      />
+      <div v-if="!input.name">
+        <a
+          v-for="plugin in plugins"
+          :key="plugin.key"
+          @click="select(plugin.key)"
+        >
+          <BlockPlugin :plugin="plugin" />
+        </a>
+        <NoResults :length="Object.keys(plugins).length" />
+      </div>
     </div>
   </UiModal>
 </template>
 
 <script>
+import { clone, filterPlugins } from '@/helpers/utils';
 import plugins from '@snapshot-labs/snapshot.js/src/plugins';
-import { clone } from '@/helpers/utils';
+
+const defaultParams = {};
 
 export default {
-  props: ['open', 'modelValue', 'space', 'proposal'],
-  emits: ['close', 'update:modelValue'],
+  props: ['open', 'plugin'],
+  emits: ['add', 'close'],
   data() {
     return {
-      plugins: [],
-      selected: false,
-      form: {}
+      input: {
+        name: '',
+        params: JSON.stringify(defaultParams, null, 2)
+      },
+      searchInput: ''
     };
   },
   watch: {
     open() {
-      if (this.modelValue && this.open) this.form = clone(this.modelValue);
-      this.selected = false;
+      if (this.plugin?.name) {
+        const plugin = this.plugin;
+        plugin.params = JSON.stringify(plugin.params, null, 2);
+        this.input = this.plugin;
+      } else {
+        this.input = {
+          name: '',
+          params: JSON.stringify(defaultParams, null, 2)
+        };
+      }
     }
   },
-  created() {
-    if (!this.space.plugins) return;
-    this.plugins = Object.fromEntries(
-      Object.keys(this.space.plugins).map(plugin => {
-        const instance = new plugins[plugin]();
-        return [plugin, instance];
-      })
-    );
+  computed: {
+    plugins() {
+      return filterPlugins(plugins, this.app.spaces, this.searchInput);
+    },
+    isValid() {
+      try {
+        const params = JSON.parse(this.input.params);
+        return !!params;
+      } catch (e) {
+        return false;
+      }
+    }
   },
   methods: {
-    getLogoUrl(plugin) {
-      return `https://raw.githubusercontent.com/snapshot-labs/snapshot.js/master/src/plugins/${plugin}/logo.png`;
+    select(plugin) {
+      this.input.name = plugin;
+    },
+    handleSubmit() {
+      const plugin = clone(this.input);
+      plugin.params = JSON.parse(plugin.params);
+      this.$emit('add', plugin);
+      this.$emit('close');
     }
   }
 };
